requirements.txt
psycopg2==2.9.9
apache-airflow==2.8.0
pandas==2.1.3
